+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Added support for configurable host url in purchase request when in test mode.
+- Added support for configurable host url in purchase request in test mode.
 
 ## [4.1.2]